+++ crates/compiler/src/driver_constants.rs
@@ -62,7 +62,7 @@ pub mod addresses {
     );
 
     // MUST match `audio-driver/src/common_memmap.wiz`
-    pub const COMMON_DATA: u16 = 0xA00 - 20;
+    pub const COMMON_DATA: u16 = 0xA00 - 24;
 
     const _: () = assert!(
         _symbols::_LAST_LOADER_SYMBOL < DRIVER_CODE,
@@ -111,15 +111,16 @@ pub const MAX_DIR_ITEMS: usize = u8::MAX as usize;
 pub const MAX_INSTRUMENTS_AND_SAMPLES: usize = u8::MAX as usize;
 pub const MAX_SOUND_EFFECTS: usize = 254;
 
-pub const PITCH_TABLE_SIZE: usize = 256;
+pub const MAX_N_PITCHES: usize = 256;
 
-pub const COMMON_DATA_POINTERS_SIZE: usize = 8 * 2;
-pub const COMMON_DATA_HEADER_SIZE: usize = COMMON_DATA_POINTERS_SIZE + 4 + (2 * PITCH_TABLE_SIZE);
+pub const COMMON_DATA_POINTERS_SIZE: usize = 10 * 2;
+pub const COMMON_DATA_HEADER_SIZE: usize = COMMON_DATA_POINTERS_SIZE + 4;
 
 pub const COMMON_DATA_PITCH_TABLE_OFFSET: usize = 20;
 pub const COMMON_DATA_DIR_TABLE_OFFSET: usize = COMMON_DATA_HEADER_SIZE;
 
 pub const COMMON_DATA_BYTES_PER_DIR: usize = 4;
+pub const COMMON_DATA_BYTES_PER_PITCH: usize = 2;
 pub const COMMON_DATA_BYTES_PER_INSTRUMENT: usize = 4;
 pub const COMMON_DATA_BYTES_PER_SOUND_EFFECT: usize = 4;